#35 - Refactor AbstractModel
